Android Testing: Add android_binary build targets for macOS and Windows (#267)
Testing isn't supported for macOS and Windows yet, but we can build the APKs.

`bazel build //ui/...` doesn't work because `android_device` targets do not work. It'll be great if we can have a `build_targets_query`/`test_targets_query` field in the yaml config to query for targets, but this will do.