Fix #309: Always clone without --recurse_submodules (#315)
It seems like when using "git clone --recurse-submodules --reference", git tries to be smart about the reference location of a submodule and computes it as "original reference + '/modules/' + repo name". However, this does the completely wrong thing on our system, because the path that git calculates doesn't exist.

I couldn't find a way to disable this behavior, but an easy fix is to just clone without --recurse_submodules and then let the existing code deal with it.